Ensure the example DAGs are all working
Some of the example DAGs are not working and some are a repeat of
what we have in another DAG. This PR aims to update the examples
so we can only have working DAGs in the examples

Update airflow/example_dags/example_kubernetes_executor.py


fixup! Update airflow/example_dags/example_kubernetes_executor.py

add back dag_decorator example

Add working template file and information why changing base image fails

Move the example dags to docs

Duplicate dag then remove non working tasks

fixup! Duplicate dag then remove non working tasks

add back path to boring cyborg

revert few changes

combine the examples

remove mistakenly included file
